catch all addition

+++ database/migrations/2026_06_04_020518_drop_email_from_providers_table.php
@@ -4,12 +4,13 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up()
     {
         Schema::table('providers', function (Blueprint $table) {
-            $table->dropColumn('email');
+            if (Schema::hasColumn('providers', 'email')) {
+                $table->dropColumn('email');
+            }
         });
     }
